Update table_interface.h
Fixed undefined symbols issue by adding four methods to table_interface.h that were originally only present in the derived class table.h. I'm not sure why this is an issue - can't a derived class have its own unique methods? - but this seems to have fixed the issue.

If there's weird behavior down the line, this would be a good place to check.