Update testing_interface_bst.cc (fixed loop bug)
Previously, the program would loop infinitely after createDestinationLoop() was entered. This was caused by usage of "cin.get()" in that method, so these calls were replaced with "cin.getline()".

If there are issues with the dynamic character arrays of destinations anywhere else in the program, this method may be the culpret...